% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\alias{assoc,multiGPATree-method}
\title{Association mapping}
\usage{
\S4method{assoc}{multiGPATree}(object, FDR = 0.05, fdrControl = "global")
}
\arguments{
\item{object}{An object of class multiGPATree.}

\item{FDR}{FDR level. Value has to be between 0 and 1.}

\item{fdrControl}{Method to control FDR. Possible values are "global" (global FDR control) and "local" (local FDR control).}
}
\value{
Returns a MX4 matrix where the row represents SNPs, the fist column indicates the association between each SNP and phenotype 1, and second column indicates the leaf in which the SNP falls
}
\description{
This function will implement association mapping for the multiGPATree model.
}
\examples{
\dontrun{
library(multiGPATree)

# load multiGPATree example data
data(simdata)

#fitting the GPATree model
fit <- multiGPATree(simdata$gwasPval, simdata$annMat)

# association mapping for the multiGPATree model fit
assoc.fit <- assoc(fit, FDR = 0.05, fdrControl = "global")
}
}
\author{
Aastha Khatiwada
}
