% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{Generate simulation data for multiple phenotypes}
\usage{
data_simulation(
  M,
  nGWAS,
  nAnn,
  percent_ones_ann,
  percent_overlap_ann,
  percent_overlap_gwas,
  trueAlphaVec
)
}
\arguments{
\item{M}{a numeric value that specifies the number of SNPs to be included in the simulation data.}

\item{nGWAS}{a numeric value that specifies the number of GWAS phenotypes to be included in the simulation.}

\item{nAnn}{a numeric value that specifies the number of functional annotations to be included in the simulation data.}

\item{percent_ones_ann}{a numeric value between 0 and 1 that specifies the percent of annotated SNPs for all functional annotations.}

\item{percent_overlap_ann}{a numeric value between 0 and 1 that specifies the percent of overlap between annotated SNPs for A1-A2 and A3-A4 and A5-A6.}

\item{percent_overlap_gwas}{a numeric value between 0 and 1 that specifies the percent of overlap P1 and P2.}

\item{trueAlphaVec}{a numeric vector between 0 and 1 to be used as the shape parameter for Beta distribution for GWAS phenotype p-values.}
}
\value{
A list containing the annotation dataframe, annMat, a dataframe containing GWAS association p-values, gwasPval.
}
\description{
This function will simulate annotation data and p-values for multiple phenotypes to apply the multGPATree method.
}
\author{
Aastha Khatiwada
}
