% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decTree.R
\name{decTree}
\alias{decTree}
\alias{decTree,multiGPATree-method}
\title{multiGPATree selected decision tree}
\usage{
\S4method{decTree}{multiGPATree}(object)
}
\arguments{
\item{object}{An object of class multiGPATree.}
}
\value{
A list containing variables in combinations selected by the multivariate decision tree (CART_PIs), the combination (CART_PIs_comb) and the predicted proportions for the selected PIs(assoc_pred).
}
\description{
This function will extract the combinations of functional annotations selected by the multivariate decision tree in the stage 2 of multiGPATree method.
}
\author{
Aastha Khatiwada
}
