% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{leaf}
\alias{leaf}
\alias{leaf,multiGPATree-method}
\title{Functional annotation tree returned by the multiGPATree model.}
\usage{
\S4method{leaf}{multiGPATree}(object)
}
\arguments{
\item{object}{An object of class multiGPATree.}
}
\value{
Returns a matrix where each row corresponds to a leaf from the multiGPATree model fit and contains information regarding the local FDR for SNPs in the leaf, and also information regarding annotations that are enriched (1) or not enriched (0) for the leaf.
}
\description{
This function will provide the annotation combinations relevant to risk-associated SNPs.
}
\examples{
\dontrun{
library(multiGPATree)

# load multiGPATree example data
data(simdata)

#fitting the multiGPATree model
fit <- multiGPATree(simdata$gwasPval, simdata$annMat)
leaf(fit)
}
}
\author{
Aastha Khatiwada
}
