% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mGPATree_s1.R
\name{mGPATree_s1}
\alias{mGPATree_s1}
\title{Implement the Stage 1 of the multiGPATree. Method}
\usage{
mGPATree_s1(gwasPval, annMat, initAlpha)
}
\arguments{
\item{gwasPval}{A matrix of M X 2 dimension where M is the number of SNPs. The first column contains the SNP id and is labeled 'SNPid' and the second column contains the GWAS association p-values and is called P1. Values in P1 must be between 0 and 1.}

\item{annMat}{A matrix of binary annotations, where row and column correspond to SNPs and annotations, respectively.}

\item{initAlpha}{Initial value for alpha estimate. Default is 0.1.}
}
\value{
This function returns a \code{List} including:
\itemize{
\item num_iter_stage1: number of iterations taken for Stage 1 of multiGPATree. to converge.
\item pi_stage1: predicted posterior probability of being a non-null SNP in Stage 1 of multiGPATree. Method.
\item alpha_stage1: estimated alpha of multiGPATree. Method.
\item beta_stage1: beta parameters from the linear model fitted at convergence of Stage 1 of multiGPATree. Method.
}
}
\description{
This function will implement the stage 1 of the multiGPATree. method.
}
\author{
Aastha Khatiwada
}
