% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mGPATree_s2.R
\name{mGPATree_s2}
\alias{mGPATree_s2}
\title{Implement the multiGPATree Method for multiple phenotypes}
\usage{
mGPATree_s2(gwasPval, annMat, alphaStage1, initPi, cpTry)
}
\arguments{
\item{gwasPval}{A matrix of M X D dimension, where M is the number of SNPs and D is the number of phenotypes. The columns contains the GWAS association p-values for the respective phenotypes. The pvalues must be between 0 and 1.}

\item{annMat}{A matrix of binary annotations, where row and column correspond to SNPs and annotations, respectively.}

\item{alphaStage1}{Alpha estimated in stage 1}

\item{initPi}{final alpha at convergence of stage 1}

\item{cpTry}{Complexity parameter (cp) value to be used to build multivaraite CART model. cpTry can be between 0 and 1 or NULL. Default is 0.001. When cpTry is NULL, multiGPATree will select the optimal cp to be used.}
}
\value{
This function returns a \code{List} including:
\itemize{
\item numIterConvergence: number of iterations taken for multiGPATree to converge.
\item alphaVec: alpha estimated by multiGPATree.
\item fit: MVPART model selected by multiGPATree.
\item fitSelectVar: annotations included in MVPART tree.
\item ziMat: posterior probability of being a non-null SNP.
\item piMat: predicted posterior probability of of being a non-null SNP.
\item licVec: list of incomplete log likelihood for all iteration of multiGPATree Method.
\item lcVec: list of complete log likelihood for all iteration of multiGPATree Method.
}
}
\description{
This function will implement the multiGPATree method for multiple phenotypes while leveraging pleiotropy.
}
\author{
Aastha Khatiwada
}
