% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGPATree_package.R
\docType{package}
\name{multiGPATree-package}
\alias{multiGPATree-package}
\alias{_PACKAGE}
\title{multiGPATree: A package to implement the multiGPATree method}
\description{
This package provides the functionality for implementing the multiGPATree method for post-GWAS analysis.
}
\author{
Aastha Khatiwada \email{asthakhatiwada@gmail.com}
}
