% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGPATree.R
\name{multiGPATree}
\alias{multiGPATree}
\title{Implement the multiGPATree Method}
\usage{
multiGPATree(gwasPval, annMat, initAlpha = 0.1, cpTry = 0.001, ncore = 2)
}
\arguments{
\item{gwasPval}{A matrix of M X D dimension, where M is the number of SNPs and D is the number of phenotypes. The columns contains the GWAS association p-values for 1, .., D phenotypes. P-values must be between 0 and 1.}

\item{annMat}{A matrix of binary annotations, where row and column correspond to SNPs and annotations, respectively.}

\item{initAlpha}{Initial value for alpha estimate. Default is 0.1.}

\item{cpTry}{Complexity parameter (cp) value to be used. cpTry can be between 0 and 1 or NULL. Default is 0.001. When cpTry is NULL, multiGPATree will select the optimal cp to be used.}

\item{ncore}{number of cores to use. we recommend changing this if you are analyzing more than one GWAS pair}
}
\value{
Contructs a multiGPATree class object
}
\description{
This function will implement Stage 1 and 2 of the multiGPATree. method.
}
\details{
The multiGPATree() function fits the multiGPATree model. It requires atleast two GWAS and expects users to provide GWAS p-value to gwasPval and binary annotation data to annMat.
It is assumed that number of rows of matrix in gwasPval and annMat are equal and correspond to the same SNP.

The assoc() function implements association mapping.

The plot() function takes in an object of class multiGPATree and will plot the functional annotation tree from the multiGPATree model.

The leaf() function takes in an object of class multiGPATree and will provide information regarding the functional annotations that are enriched (1) or not enriched (0) for SNPs in any leaf of the multiGPATree model plot.

The prune() function takes in an object of class multiGPATree and a cp parameter and will prune the multiGPATree model result. This function can be useful when the tree obtained from multiGPATree model is larger.
}
\examples{

\dontrun{
library(multiGPATree)

# load multiGPATree example data
data(simdata)

# fitting the multiGPATree model
fit <- multiGPATree(simdata$gwasPval, simdata$annMat)

# get functional annotation information
leaf(fit)

# association mapping
assoc.mgpatree <- assoc(fit, FDR = 0.01, fdrControl = 'global')

# pruning the multiGPATree model fit
pruned.fit <- prune(fit, cp = 0.005)

# plotting the multiGPATree model results
plot(fit)
plot(pruned.fit)
}
}
\author{
Aastha Khatiwada
}
