% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,multiGPATree,missing-method}
\title{Plot the functional annotation tree obtained from the multiGPATree model fit.}
\usage{
\S4method{plot}{multiGPATree,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class multiGPATree.}

\item{y}{missing (not required).}

\item{...}{...}
}
\value{
Returns a plot for the functional annotation tree from the multiGPATree model fit.
}
\description{
This function will plot the functional annotation tree for the multiGPATree model fit.
}
\examples{
\dontrun{
library(multiGPATree)

# load multiGPATree example data
data(simdata)

#fitting the multiGPATree model
fit <- multiGPATree(simdata$gwasPval, simdata$annMat)

# plotting the multiGPATree model fit
plot(fit)
}
}
\author{
Aastha Khatiwada
}
