% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\alias{prune,multiGPATree-method}
\title{Prune multiGPATree model fit}
\usage{
\S4method{prune}{multiGPATree}(object, cp = 0.001)
}
\arguments{
\item{object}{An object of class multiGPATree.}

\item{cp}{The cp parameter to be used for pruning. cp must be between 0 and 1.}
}
\value{
multiGPATree model output.
}
\description{
This function will prune the multiGPATree model fit using the given cp value.
}
\examples{
\dontrun{
library(multiGPATree)

# load multiGPATree example data
data(simdata)

#fitting the multiGPATree model
fit <- multiGPATree(simdata$gwasPval, simdata$annMat)

# pruning the multiGPATree model fit
pruned.fit <- prune(fit, cp = 0.005)
}
}
\author{
Aastha Khatiwada
}
