% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.lists.R
\name{rpart.lists}
\alias{rpart.lists}
\title{Creates lists of variable values (factor levels) associated with each rule in an \pkg{rpart} object.}
\usage{
rpart.lists(object)
}
\arguments{
\item{object}{an rpart object}
}
\value{
a list of lists
}
\description{
Creates lists of variable values (factor levels) associated with each rule in an \pkg{rpart} object.
}
