% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.rules.R
\name{rpart.rules}
\alias{rpart.rules}
\title{Returns a list of strings summarizing the branch path to each node.}
\usage{
rpart.rules(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
Returns a list of strings summarizing the branch path to each node.
}
