% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.rules.table.R
\name{rpart.rules.table}
\alias{rpart.rules.table}
\title{Returns an unpivoted table of branch paths (subrules) associated with each node.}
\usage{
rpart.rules.table(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
Returns an unpivoted table of branch paths (subrules) associated with each node.
}
