% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.subrules.table.R
\name{rpart.subrules.table}
\alias{rpart.subrules.table}
\title{Returns an unpivoted table of variable values (factor levels) associated with each branch.}
\usage{
rpart.subrules.table(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
Returns an unpivoted table of variable values (factor levels) associated with each branch.
}
