% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\alias{simdata,multiGPATree-method}
\title{simdata}
\format{
A List that contains two data frames for two phenotypes/traits of interest.
\itemize{ The first data frame contains 1,000 rows and two columns. The rows represent SNPs and the columns represent the GWAS association p-value for the association between the SNPs and two phenotype
}
\itemize{ The second dataframe contains 1,000 rows and 10 columns. The rows represent SNPs and the columns represent 10 binary functional annotations.
}
}
\source{
Simulated Data
}
\usage{
simdata
}
\description{
Simulated data for the multiGPATree package
}
\keyword{datasets}
